requirements.txt
docker>=6.1.3
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
